updated build for mac, fixed issues with gcc7, also includes massive upgrade to transaction pool and fusion algos.

+++ src/CryptoNoteCore/TransactionPool.h
@@ -150,7 +150,7 @@ namespace CryptoNote {
     struct TransactionDetails : public TransactionCheckInfo {
       Crypto::Hash id;
       Transaction tx;
-      size_t blobSize;
+      uint64_t blobSize;
       uint64_t fee;
       bool keptByBlock;
       time_t receiveTime;
